# students/YarmolaAleksandr/task_02/src/apps/server/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Longread Blog API",
    "version": "1.0.0",
    "description": "API для блога: пользователи, посты, теги, комментарии, лайки"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Локальный сервер"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "email": { "type": "string" },
          "name": { "type": "string" },
          "role": { "type": "string" }
        }
      },
...
      "Tag": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" }
        }
      },
...
    }
  },
  "security": [
    { "BearerAuth": [] }
  ],
  "paths": {
...
    "/tags": {
      "get": {
        "tags": ["tags"],
        "summary": "Список тегов",
        "responses": {
          "200": { "description": "Теги", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Tag" } } } } }
        }
      },
      "post": {
        "tags": ["tags"],
        "summary": "Создать тег",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "name": { "type": "string" } } }
            }
          }
        },
        "responses": {
          "200": { "description": "Тег создан" },
          "401": { "description": "Нет доступа" }
        }
      }
    },
    "/tags/{id}": {
      "get": {
        "tags": ["tags"],
        "summary": "Получить тег по id",
        "parameters": [{
          "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }
        }],
        "responses": {
          "200": { "description": "Тег", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Tag" } } } },
          "404": { "description": "Не найдено" }
        }
      },
      "put": {
        "tags": ["tags"],
        "summary": "Обновить тег",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{
          "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }
        }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "name": { "type": "string" } } }
            }
          }
        },
        "responses": {
          "200": { "description": "Тег обновлён" }
        }
      },
      "delete": {
        "tags": ["tags"],
        "summary": "Удалить тег",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{
          "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }
        }],
        "responses": {
          "200": { "description": "Тег удалён" }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": ["users"],
        "summary": "Список пользователей (только admin)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": { 
            "description": "Список пользователей", 
            "content": { 
              "application/json": { 
                "schema": { 
                  "type": "object",
                  "properties": {
                    "users": { "type": "array", "items": { "$ref": "#/components/schemas/User" } },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "page": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "totalPages": { "type": "integer" }
                      }
                    }
                  }
                }
              } 
            } 
          },
          "401": { "description": "Требуется авторизация" },
          "403": { "description": "Требуются права администратора" }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["users"],
        "summary": "Получить пользователя по ID (только admin)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{
          "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }
        }],
        "responses": {
          "200": { "description": "Данные пользователя", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "404": { "description": "Пользователь не найден" },
          "403": { "description": "Требуются права администратора" }
        }
      },
      "put": {
        "tags": ["users"],
        "summary": "Обновить пользователя (только admin)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{
          "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }
        }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "name": { "type": "string" },
                  "role": { "type": "string", "enum": ["user", "admin"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Пользователь обновлён", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "400": { "description": "Ошибка валидации" },
          "403": { "description": "Требуются права администратора" }
        }
      },
      "delete": {
        "tags": ["users"],
        "summary": "Удалить пользователя (только admin)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{
          "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }
        }],
        "responses": {
          "200": { "description": "Пользователь удалён" },
          "400": { "description": "Нельзя удалить самого себя" },
          "403": { "description": "Требуются права администратора" },
          "404": { "description": "Пользователь не найден" }
        }
      }
    }
  }
}